requirements.txt
asgiref==3.5.2
astroid==2.11.7
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
coreapi==2.3.3
coreschema==0.0.4
dependency-injector==4.39.1
dill==0.3.5.1
Django==4.0.6
djangorestframework==3.13.1
drf-yasg==1.21.3
h11==0.13.0
idna==3.3
inflection==0.5.1
isort==5.10.1
itypes==1.2.0
Jinja2==3.1.2
lazy-object-proxy==1.7.1
MarkupSafe==2.1.1
mccabe==0.7.0
mediatr==1.3.2
packaging==21.3
platformdirs==2.5.2
psycopg2-binary==2.9.3
pydantic==1.9.1
pylint==2.14.5
pyparsing==3.0.9
pytz==2022.1
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.2
tomli==2.0.1
tomlkit==0.11.1
typing_extensions==4.3.0
uritemplate==4.1.1
urllib3==1.26.11
uvicorn==0.18.2
wrapt==1.14.1
